requirements.txt

amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
# click==8.2.1
# click-didyoumean==0.3.1
# click-plugins==1.1.1.2
# click-repl==0.3.0
# colorama==0.4.6
# crispy-bootstrap5==2025.6
# cron-descriptor==1.4.5
# Django==5.2.4
# django-celery-beat==2.8.1
django-ckeditor==6.5.1

# django-crispy-forms==2.4
# django-timezone-field==7.1
# kombu==5.5.4
# packaging==25.0
pillow==11.3.0
# prompt_toolkit==3.0.51
# python-crontab==3.3.0
# python-dateutil==2.9.0.post0
# python-decouple==3.8
# qrcode==8.2
# redis==6.2.0
# six==1.17.0
# sqlparse==0.5.3
# tzdata==2025.2
# vine==5.1.0
# wcwidth==0.2.13

django-taggit
